Move the Web table of contents code to its own module and convert to TypeScript
